Storing BPM, made it work with milliseconds.

+++ app/src/main/java/com/example/andersson/musicapp/Instrument/AbstractInstrumentThread.java
@@ -100,7 +100,7 @@ public abstract class AbstractInstrumentThread extends Thread implements Observe
     public void setLoopTime(double loopTime) {
 
         this.loopTime = loopTime;
-        ((AbstractInstrumentActivity) activity).setLoopTime((int) loopTime);
+        ((AbstractInstrumentActivity) activity).setLoopTime((double) loopTime);
 
     }
